test : Baseball 생성시, 숫자 야구공의 숫자와 위치가 제대로 생성되는지 확인하는 테스트 코드 작성
